+++ cmd/template/cluster/runner.go
@@ -74,16 +74,6 @@ func (r *runner) run(ctx context.Context, client k8sclient.Interface) error {
 	}
 
 	switch r.flag.Provider {
-	case key.ProviderAWS:
-		err = provider.WriteAWSTemplate(ctx, client, output, config)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-	case key.ProviderAzure:
-		err = provider.WriteAzureTemplate(ctx, client, output, config)
-		if err != nil {
-			return microerror.Mask(err)
-		}
 	case key.ProviderCAPA:
 		err = provider.WriteCAPATemplate(ctx, client, output, config)
 		if err != nil {
@@ -163,9 +153,7 @@ func (r *runner) getClusterConfig() (common.ClusterConfig, error) {
 		return common.ClusterConfig{}, microerror.Mask(err)
 	}
 
-	if r.flag.Provider != key.ProviderAWS {
-		config.Namespace = key.OrganizationNamespaceFromName(config.Organization)
-	}
+	config.Namespace = key.OrganizationNamespaceFromName(config.Organization)
 
 	return config, nil
 }