+++ internal/key/templates.go
@@ -21,24 +21,6 @@ const CatalogCRTemplate = `
 {{ .CatalogCR -}}
 `
 
-const ClusterAWSCRsTemplate = `
-{{- .ClusterCR -}}
----
-{{ .AWSClusterCR -}}
----
-{{ .G8sControlPlaneCR -}}
----
-{{ .AWSControlPlaneCR -}}
-`
-
-const ClusterAzureCRsTemplate = `
-{{- .ProviderClusterCR -}}
----
-{{ .ClusterCR -}}
----
-{{ .MasterMachineCR -}}
-`
-
 const ClusterCAPACRsTemplate = `
 {{- .ClusterCR -}}
 ---
@@ -63,10 +45,6 @@ const ClusterEKSCRsTemplate = `
 {{ .AWSManagedControlPlaneCR -}}
 ---
 {{ .AWSClusterRoleIdentityCR -}}
-`
-
-const NetworkPoolCRsTemplate = `
-{{- .NetworkPoolCR -}}
 `
 
 const MachinePoolAWSCRsTemplate = `